requirements.txt
kafka-python
pyyaml
tornado==3.2.1